% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vl.R
\name{plot_viral_load_trajectory}
\alias{plot_viral_load_trajectory}
\title{Plot viral load trajectory for one individual}
\usage{
plot_viral_load_trajectory(age_group = NULL, params, spaghetti = FALSE, N = 50)
}
\arguments{
\item{age_group}{Index of the age group (starting at 0). If NULL, one is sampled.}

\item{params}{List of parameters (must include dt, VL_days, k, gamma, sigma, age_groups)}

\item{I0}{Initial viral load (default 0.1)}

\item{baseline}{Baseline viral load (default 0.0)}
}
\description{
Plot viral load trajectory for one individual
}
